final version of week 5

+++ app/Http/Controllers/ReceipeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Receipe;
+use App\Category;
 use Illuminate\Http\Request;
 
 class ReceipeController extends Controller
@@ -12,9 +13,15 @@ class ReceipeController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $data = Receipe::all();
+        $data = Receipe::where('author_id', auth()->id())->get();
         return view('home', compact('data'));
     }
 
@@ -25,7 +32,8 @@ class ReceipeController extends Controller
      */
     public function create()
     {
-        return view('create');
+        $category = Category::all();
+        return view('create', compact('category'));
     }
 
     /**
@@ -40,6 +48,7 @@ class ReceipeController extends Controller
             'name' => 'required',
             'ingredients' => 'required',
             'category' => 'required',
+            // 'author_id' => 'required'
         ]);
 
       /*$receipe = new Receipe();
@@ -51,7 +60,7 @@ class ReceipeController extends Controller
         $receipe->save(); */
 
     //Use for fillable in Model
-        Receipe::create($validatedData);
+        Receipe::create($validatedData + ['author_id' => auth()->id()]);
 
       //Use for guraded in Model
      /* Receipe::create([
@@ -71,6 +80,10 @@ class ReceipeController extends Controller
      */
     public function show(Receipe $receipe)
     {
+        // if ($receipe->author_id != auth()->id()) {
+        //     abort(403);
+        // }
+        $this->authorize('view',$receipe);
         return view('show', compact('receipe'));
     }
 
@@ -82,7 +95,10 @@ class ReceipeController extends Controller
      */
     public function edit(Receipe $receipe)
     {
-        return view('edit', compact('receipe'));
+        $this->authorize('view',$receipe);
+
+        $category = Category::all();
+        return view('edit', compact('receipe', 'category'));
     }
 
     /**
@@ -94,6 +110,8 @@ class ReceipeController extends Controller
      */
     public function update(Receipe $receipe)
     {
+        $this->authorize('view',$receipe);
+
         $validatedData = request()->validate([
             'name' => 'required',
             'ingredients' => 'required',
@@ -113,6 +131,8 @@ class ReceipeController extends Controller
      */
     public function destroy(Receipe $receipe)
     {
+        $this->authorize('view',$receipe);
+
         $receipe->delete();
         return redirect('receipe');
     }
